requirements.txt
absl-py==0.13.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
gevent==21.1.2
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
rdkit-pypi==2021.3.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.0
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
torch==1.9.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
